// somdata/types/classes.ts
export enum SomClass {
  MAGE = "Mage",

  DRAGOON = "Dragoon",
  NINJA = "Ninja",
  PALADIN = "Paladin",
  MONK = "Monk",
  ASSASSIN = "Assassin",
  KNIGHT = "Knight",
  VAMPIRE_HUNTER = "Vampire Hunter",
  PIRATE = "Pirate",
  SKUD = "Skud",
  ANGEL = "Angel (f)",
  VALKYRIE = "Valkyrie (f)",
}

export const SomClassesMelee: SomClass[] = [
  SomClass.DRAGOON,
  SomClass.NINJA,
  SomClass.PALADIN,
  SomClass.MONK,
  SomClass.ASSASSIN,
  SomClass.KNIGHT,
  SomClass.VAMPIRE_HUNTER,
  SomClass.PIRATE,
  SomClass.ANGEL,
  SomClass.VALKYRIE
];